Working on searching a topsig index (not working yet)
git-svn-id: https://groke.otago.ac.nz/ant/ant@1163 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/ranking_function_pregen.h
@@ -17,10 +17,11 @@ class ANT_ranking_function_pregen : public ANT_ranking_function
 {
 private:
 	ANT_pregen *pregen;
-	int ascending;
+	long ascending;
+
 public:
-	ANT_ranking_function_pregen(ANT_search_engine *engine, ANT_pregen *pregen, int ascending) : ANT_ranking_function(engine), pregen(pregen), ascending(ascending) {};
-	ANT_ranking_function_pregen(long long documents, ANT_compressable_integer *document_lengths, int ascending) : ANT_ranking_function(documents, document_lengths), pregen(NULL), ascending(ascending) {}
+	ANT_ranking_function_pregen(ANT_search_engine *engine, ANT_pregen *pregen, long ascending) : ANT_ranking_function(engine), pregen(pregen), ascending(ascending) {};
+	ANT_ranking_function_pregen(long long documents, ANT_compressable_integer *document_lengths, long ascending) : ANT_ranking_function(documents, document_lengths), pregen(NULL), ascending(ascending) {}
 	virtual ~ANT_ranking_function_pregen() {}
 
 	virtual void relevance_rank_top_k(ANT_search_engine_result *accumulator, ANT_search_engine_btree_leaf *term_details, ANT_compressable_integer *impact_ordering, long long trim_point, double prescalar, double postscalar);
